style: 랜딩 페이지 디렉토리 구조 변경
