+++ client/src/pages/MainPage/Nav/Nav.style.ts
@@ -1,25 +1,8 @@
 import {css} from '@emotion/react';
 
-import {Theme} from '@components/Design/theme/theme.type';
-
-export const navStyle = (theme: Theme) =>
-  css({
-    position: 'fixed',
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    padding: '1rem',
-
-    top: '0',
-    width: '100%',
-    maxWidth: '768px',
-    zIndex: theme.zIndex.navBackgroundColor,
-    height: '4rem',
-    backgroundColor: 'white',
-  });
-
-export const logoStyle = css({
+export const navStyle = css({
   display: 'flex',
-  gap: '0.5rem',
+  justifyContent: 'space-between',
   alignItems: 'center',
+  height: '2.5rem',
 });
