+++ client/src/pages/MainPage/Section/DescriptionSection/index.ts
@@ -1,0 +1,1 @@
+export {default as DescriptionSection} from './DescriptionSection';
