+++ client/src/pages/MainPage/Section/FeatureSection/Calculate/index.ts
@@ -1,0 +1,1 @@
+export {default as Calculate} from './Calculate';
