+++ client/src/pages/MainPage/Section/FeatureSection/CheckDeposit/CheckDeposit.tsx
@@ -1,0 +1,25 @@
+import {Text} from '@components/Design';
+
+import {backgroundStyle, imageStyle, sectionStyle, textContainerStyle} from './CheckDeposit.style';
+
+const CheckDeposit = () => {
+  return (
+    <div css={backgroundStyle}>
+      <div css={sectionStyle}>
+        <div css={textContainerStyle}>
+          <Text size="subTitle" responsive={true}>
+            입금 확인도 걱정 없어요
+          </Text>
+          <Text size="body" responsive={true}>
+            {`매번 누가 입금했는지 따로 기억할 필요 없어요.
+            참여자 관리 메뉴를 통해 더 이상 놓치는 사람 없이
+            간편하게 관리할 수 있어요.`}
+          </Text>
+        </div>
+        <img src={`${process.env.IMAGE_URL}/feature3.svg`} css={imageStyle} />
+      </div>
+    </div>
+  );
+};
+
+export default CheckDeposit;
