+++ client/src/pages/MainPage/Section/FeatureSection/CheckDeposit/index.ts
@@ -1,0 +1,1 @@
+export {default as CheckDeposit} from './CheckDeposit';
