+++ client/src/pages/MainPage/Section/FeatureSection/Photo/index.ts
@@ -1,0 +1,1 @@
+export {default as Photo} from './Photo';
