+++ client/src/pages/MainPage/Section/FeatureSection/SimpleShare/index.ts
@@ -1,0 +1,1 @@
+export {default as SimpleShare} from './SimpleShare';
