+++ client/src/pages/MainPage/Section/FeatureSection/SimpleTransfer/index.ts
@@ -1,0 +1,1 @@
+export {default as SimpleTransfer} from './SimpleTransfer';
