+++ client/src/pages/MainPage/Section/FeatureSection/index.ts
@@ -1,0 +1,1 @@
+export {default as FeatureSection} from './FeatureSection';
