+++ client/src/pages/MainPage/Section/MainSection/index.ts
@@ -1,0 +1,1 @@
+export {default as MainSection} from './MainSection';
