+++ client/src/pages/MainPage/Section/ReportSection/ReportSection.style.ts
@@ -1,0 +1,17 @@
+import {css} from '@emotion/react';
+
+export const reportSectionStyle = css({
+  display: 'flex',
+  flexDirection: 'column',
+  justifyContent: 'center',
+  padding: '3rem',
+  gap: '2rem',
+  width: '100%',
+  backgroundColor: 'white',
+});
+
+export const textContainerStyle = css({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '1rem',
+});
