+++ client/src/pages/MainPage/Section/ReportSection/ReportSection.tsx
@@ -1,21 +1,11 @@
-import {css} from '@emotion/react';
-
 import Text from '@components/Design/components/Text/Text';
+
+import {reportSectionStyle, textContainerStyle} from './ReportSection.style';
 
 const ReportSection = () => {
   return (
-    <div
-      css={css({
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'center',
-        padding: '3rem',
-        gap: '2rem',
-        width: '100%',
-        backgroundColor: 'white',
-      })}
-    >
-      <div css={css({display: 'flex', flexDirection: 'column', gap: '1rem'})}>
+    <div css={reportSectionStyle}>
+      <div css={textContainerStyle}>
         <Text size="subTitle">친구에게 링크를 공유하세요</Text>
         <Text size="body" textColor="gray">
           {`지출내역과 인원변동을 통해
