+++ client/src/pages/MainPage/Section/ReportSection/index.ts
@@ -1,0 +1,1 @@
+export {default as ReportSection} from './ReportSection';
